age-secret/pyproject.toml
[project]
name = "encryptor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyrage>=1.2.4",
    "pyyaml>=6.0.2",
]
